fix bug with NPE

+++ app/src/main/java/ru/solandme/washwait/network/weather/darksky/model/DarkSkyForecast.java
@@ -17,7 +17,7 @@ public class DarkSkyForecast {
     private String timezone;
     @SerializedName("offset")
     @Expose
-    private int offset;
+    private double offset;
     @SerializedName("currently")
     @Expose
     private Currently currently;
@@ -58,11 +58,11 @@ public class DarkSkyForecast {
         this.timezone = timezone;
     }
 
-    public int getOffset() {
+    public double getOffset() {
         return offset;
     }
 
-    public void setOffset(int offset) {
+    public void setOffset(double offset) {
         this.offset = offset;
     }
